Factor mustache-specific template rendering into its own module; add jade-specific module stub
