+++ lib/render.js
@@ -2,11 +2,8 @@
 /*jslint node: true */
 'use strict';
 
-var partials,
-    mustache = require('mustache'),
+var templateEngine,
     fs = require('fs'),
-    path = require('path'),
-    file = require('./file'),
     codeRegExp = /<code\s+class="(html|javascript|css)">/g,
     prismLang = {
         javascript: 'javascript',
@@ -14,23 +11,24 @@ var partials,
         html: 'markup'
     };
 
-function loadPartials() {
-    partials = {};
+function render(template, data, meta) {
+//console.log(JSON.stringify({template:template, data:data, meta:meta}));
+    var module;
+    if (!templateEngine) {
+        switch (meta.data.templateEngine && meta.data.templateEngine.toLowerCase()) {
+            case "jade":
+                module = "jade";
+                break;
+            case "mustache":
+            default:
+                module = "mustache";
+        }
 
-    var partialTemplatePath = path.join(process.cwd(), 'templates', 'partials');
-
-    //Load partials
-    fs.readdirSync(partialTemplatePath).forEach(function (fileName) {
-        partials[path.basename(fileName, path.extname(fileName))] = file.read(path.join(partialTemplatePath, fileName));
-    });
-}
-
-function render(template, data) {
-    if (!partials) {
-        loadPartials();
+        templateEngine = require("./render/" + module);
     }
+
     //Convert to HTML, and change any code blocks to use language- for prism.
-    return mustache.render(template, data, partials).replace(codeRegExp, function (match, lang) {
+    return templateEngine(template, data).replace(codeRegExp, function (match, lang) {
         return '<code class="language-' + prismLang[lang] + '">';
     });
 }
